App: Let the activity handles model
This patch makes the activity to handle model.
When a model is added, it can be accessed through an activity.
